+++ sdk/@launchdarkly/react-native-ld-session-replay/ios/SessionReplayReactNative.h
@@ -1,4 +1,11 @@
+#ifdef RCT_NEW_ARCH_ENABLED
 #import <SessionReplayReactNativeSpec/SessionReplayReactNativeSpec.h>
 
 @interface SessionReplayReactNative : NSObject <NativeSessionReplayReactNativeSpec>
 @end
+#else
+#import <React/RCTBridgeModule.h>
+
+@interface SessionReplayReactNative : NSObject <RCTBridgeModule>
+@end
+#endif